refactor(types): centralize shared interfaces
