+++ src/adminPanel/types.ts
@@ -1,77 +1,18 @@
-export  interface User {
-  id: string;
-  username: string;
-  email: string;
+import {
+  User as SharedUser,
+  Club,
+  Title,
+  Player,
+  PlayerAttributes,
+  PlayerContract,
+} from '../types/shared';
+
+export { Club, Title, Player, PlayerAttributes, PlayerContract } from '../types/shared';
+
+export interface User extends SharedUser {
   role: 'admin' | 'dt' | 'user';
   status: 'active' | 'inactive';
   createdAt: string;
-  clubId?: string;
-}
-
-export interface Club {
-  id: string;
-  name: string;
-  slug: string;
-  logo: string;
-  foundedYear: number;
-  stadium: string;
-  budget: number;
-  manager: string;
-  managerId?: string;
-  playStyle: string;
-  primaryColor: string;
-  secondaryColor: string;
-  description: string;
-  titles: Title[];
-  reputation: number;
-  fanBase: number;
-  morale: number;
-  createdAt?: string;
-}
-
-export interface Title {
-  id: string;
-  name: string;
-  year: number;
-  type: 'league' | 'cup' | 'supercup' | 'other';
-}
-
-export interface Player {
-  id: string;
-  name: string;
-  age: number;
-  position: string;
-  nationality: string;
-  dorsal: number;
-  clubId: string;
-  overall: number;
-  potential: number;
-  transferListed: boolean;
-  matches: number;
-  transferValue: number;
-  value: number;
-  image: string;
-  attributes: PlayerAttributes;
-  contract: PlayerContract;
-  form: number;
-  goals: number;
-  assists: number;
-  appearances: number;
-  price?: number;
-}
-
-export interface PlayerAttributes {
-  pace: number;
-  shooting: number;
-  passing: number;
-  dribbling: number;
-  defending: number;
-  physical: number;
-}
-
-export interface PlayerContract {
-  expires: string;
-  salary: number;
 }
 
 export interface Tournament {
