+++ src/components/tacticas/SetPiecesManager.tsx
@@ -1,4 +1,4 @@
-import { Player } from '../../types';
+import { Player } from '../../types/shared';
 
 interface SetPieces {
   penalties?: string;
