+++ src/pages/Plantilla.tsx
@@ -4,7 +4,7 @@ import ResumenClub from '../components/plantilla/ResumenClub';
 import PlayerTable from '../components/plantilla/PlayerTable';
 import { useAuthStore } from '../store/authStore';
 import { useDataStore } from '../store/dataStore';
-import { Player as FullPlayer } from '../types';
+import { Player as FullPlayer } from '../types/shared';
 
 const PlayerDrawer = lazy(() => import('../components/plantilla/PlayerDrawer'));
